requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
dnspython==2.4.2
cryptography==41.0.7
python-dotenv==1.0.0
django-cors-headers==4.3.1
# convex==0.0.8
